+++ ababot/bot/src/background_tasks/abakus/mod.rs
@@ -1,2 +1,3 @@
 mod abakus;
+mod types;
 pub use abakus::run;
